Bluetooth: Mesh: Rewrite sensor API to use raw values
This commit introduces a new API for the Mesh sensor models, where the
sensor data is kept as the raw encoded values. The old APIs based on
struct sensor_value are deprecated and toggleable using a KConfig
option.

This change is made in order to be able to accurately handle all values
of every sensor format.

Signed-off-by: Ludvig Jordet <ludvig.jordet@nordicsemi.no>